Added YOLOv5, fixed borders, changed bboxes from int to double array

+++ pytb/detection/bboxes/bboxes_2d_detector/background_subtractor/background_subtractor.py
@@ -78,5 +78,5 @@ class BackgroundSubtractor(BBoxes2DDetector):
             bboxes.append([x, y, w, h])
         end = default_timer()
 
-        return BBoxes2D(end-start, np.array(bboxes).astype(int), np.zeros(len(bboxes)).astype(int),
+        return BBoxes2D(end-start, np.array(bboxes), np.zeros(len(bboxes)).astype(int),
                         np.ones(len(bboxes)), frame.shape[1], frame.shape[0])
